New  Beginnings
Congratulations on finding this mod, however, it is well... Not even developed yet! It should be an attempt to add Pokemon to minecraft without being malware. Yes there exist mods that will IP log you, scam for money on servers, and place viruses on your pc because you want pokemon; this mod is not one of them!